~ Fixed some more bugs, the sorter is now almost working.

+++ valves.h
@@ -9,8 +9,8 @@
 #define TIME_TO_BOTTOM_OF_PICTURE ((uint32) CPU_FREQ / 1000000 * 191565)
 #define TIME_TO_VALVES ((uint32) CPU_FREQ / 1000000 * 247309)
 
-#define PIXEL_BEGIN_FIRST_VALVE 20
-#define PIXEL_END_LAST_VALVE 334
+#define PIXEL_BEGIN_FIRST_VALVE ((uint16) 20)
+#define PIXEL_END_LAST_VALVE ((uint16) 334)
 
 void valves_init();
 void valves_insertEvent(t_time const begin_time, t_time const end_time, t_index const first_valve, t_index const last_valve);
